requirements.txt
python-dotenv==1.0.1
pydantic==2.9.1
pandas==2.2.2
requests==2.32.3
tqdm==4.66.5
python-dateutil==2.9.0.post0
pytz==2024.1
streamlit==1.37.1
